requirements.txt
aioboto3==13.1.0
aiobotocore==2.13.1
aiofiles==24.1.0
aiohttp==3.9.5
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
black==24.1.1
boto3==1.34.131
botocore==1.34.131
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
diskcache==5.6.3
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.4
jmespath==1.0.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
orjson==3.10.6
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.7
requests==2.31.0
rich==13.7.1
rqdb==1.6.0
s3transfer==0.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.37.2
twilio==9.2.3
typer==0.12.3
typing_extensions==4.9.0
ujson==5.10.0
urllib3==2.0.7
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
win32-setctime==1.1.0
wrapt==1.16.0
yarl==1.9.4
